+++ api/index.ts
@@ -1,5 +1,6 @@
 export * from './account'
 export * from './blockchains'
+export * from './calendar'
 export * from './games'
 export * from './ratings'
 export * from './server'
