+++ types/calendar.d.ts
@@ -1,0 +1,10 @@
+declare namespace Calendar {
+  type Jam = {
+    summary: string
+    description: string
+    start: Date
+    end: Date
+    link: string
+    uid: string
+  }
+}
