+++ utils/index.ts
@@ -1,4 +1,5 @@
 export * from './algolia'
+export * from './calendar'
 export * from './chains'
 export * from './cookie'
 export * from './error'
@@ -46,3 +47,23 @@ export const isStringNumber = (value: string): boolean => {
 
 export const sleep = (ms: number) =>
   new Promise((resolve) => setTimeout(resolve, ms))
+
+// compose
+// https://github.com/reduxjs/redux/blob/master/src/compose.ts
+// eslint-disable-next-line @typescript-eslint/ban-types
+export const compose = (...fn: Function[]) => {
+  if (fn.length === 0) {
+    return <T>(arg: T) => arg
+  }
+
+  if (fn.length === 1) {
+    return fn[0]
+  }
+
+  return fn.reduce(
+    (a, b) =>
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      (...args: any) =>
+        a(b(...args))
+  )
+}
