+++ utils/string.ts
@@ -20,3 +20,13 @@ export const enumWords = (words: string[]) =>
 export function stringSlice(str: string, start: number, end: number): string {
   return `${str.slice(0, start)}...${str.slice(-Math.abs(end))}`
 }
+
+/**
+ * string summary
+ * @param str
+ * @param length
+ * @returns
+ */
+export function stringSummary(str: string, length: number): string {
+  return str.length >= length ? str.slice(0, length - 3) + '...' : str
+}
